Refactor: move main function to its own file
